postgres_to_es/requirements.txt
asgiref==3.5.2
attrs==22.1.0
backoff==2.2.1
certifi==2022.12.7
click==8.1.3
coverage==7.0.3
elastic-transport==8.4.0
elasticsearch7==7.17.8
eradicate==2.1.0
flake8==5.0.4
iniconfig==1.1.1
isort==5.10.1
mccabe==0.7.0
mr-proper==0.0.7
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytest==7.2.0
pytest-cov==4.0.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
stdlib-list==0.8.0
typing_extensions==4.4.0
urllib3==1.26.13
